Update mpesa cloud functions. A lot of work lost
Code implementing mpesa payment inevitably lost by git checkout
