+++ BuukRides/app/build.gradle
@@ -57,6 +57,7 @@ dependencies {
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.0'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
+    implementation 'com.loopj.android:android-async-http:1.4.9'
 
 
     implementation 'com.ibm.watson:assistant:8.1.0'
