+++ BuukRides/app/src/main/res/values/styles.xml
@@ -6,6 +6,7 @@
         <item name="colorPrimary">@color/colorPrimary</item>
         <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
         <item name="colorAccent">@color/colorAccent</item>
+<!--        <item name="android:fontFamily">@assets/Montserrat-Regular</item>-->
     </style>
 
     <style name="AppTheme.NoActionBar">
