--- notification_function/firebase.json
@@ -1,7 +1,0 @@
-{
-  "functions": {
-    "predeploy": [
-      "npm --prefix \"$RESOURCE_DIR\" run lint"
-    ]
-  }
-}
